--- packages/django-cf/django_cf/do_binding/storage.py
@@ -1,8 +1,0 @@
-storage = None
-
-def set_storage(db):
-    global storage
-    storage = db
-
-def get_storage():
-    return storage
