--- packages/django-cf/templates/d1/src/django_cf/d1_api/client.py
@@ -1,5 +1,0 @@
-from django.db.backends.sqlite3.client import DatabaseClient as SQLiteDatabaseClient
-
-
-class DatabaseClient(SQLiteDatabaseClient):
-    pass
