--- packages/django-cf/templates/d1/src/django_cf/d1_api/creation.py
@@ -1,5 +1,0 @@
-from django.db.backends.sqlite3.creation import DatabaseCreation as SQLiteDatabaseCreation
-
-
-class DatabaseCreation(SQLiteDatabaseCreation):
-    pass
