--- packages/django-cf/templates/d1/src/django_cf/d1_binding/introspection.py
@@ -1,5 +1,0 @@
-from django.db.backends.sqlite3.introspection import DatabaseIntrospection as SQLiteDatabaseIntrospection
-
-
-class DatabaseIntrospection(SQLiteDatabaseIntrospection):
-    pass
