--- packages/django-cf/templates/d1/src/django_cf/d1_binding/operations.py
@@ -1,58 +1,0 @@
-from django.db.backends.sqlite3.operations import DatabaseOperations as SQLiteDatabaseOperations
-
-
-class DatabaseOperations(SQLiteDatabaseOperations):
-    # This patches some weird bugs related to the Database class
-    def _quote_params_for_last_executed_query(self, params):
-        """
-        Only for last_executed_query! Don't use this to execute SQL queries!
-        """
-        # This function is limited both by SQLITE_LIMIT_VARIABLE_NUMBER (the
-        # number of parameters, default = 999) and SQLITE_MAX_COLUMN (the
-        # number of return values, default = 2000). Since Python's sqlite3
-        # module doesn't expose the get_limit() C API, assume the default
-        # limits are in effect and split the work in batches if needed.
-        BATCH_SIZE = 999
-        if len(params) > BATCH_SIZE:
-            results = ()
-            for index in range(0, len(params), BATCH_SIZE):
-                chunk = params[index: index + BATCH_SIZE]
-                results += self._quote_params_for_last_executed_query(chunk)
-            return results
-
-        sql = "SELECT " + ", ".join(["QUOTE(?)"] * len(params))
-        # Bypass Django's wrappers and use the underlying sqlite3 connection
-        # to avoid logging this query - it would trigger infinite recursion.
-
-        cursor = self.connection.connection.cursor()
-        # Native sqlite3 cursors cannot be used as context managers.
-        # try:
-        #     return cursor.execute(sql, params).fetchone()
-        # finally:
-        #     cursor.close()
-
-    def last_executed_query(self, cursor, sql, params):
-        # Python substitutes parameters in Modules/_sqlite/cursor.c with:
-        # bind_parameters(state, self->statement, parameters);
-        # Unfortunately there is no way to reach self->statement from Python,
-        # so we quote and substitute parameters manually.
-        if params:
-            if isinstance(params, (list, tuple)):
-                params = self._quote_params_for_last_executed_query(params)
-            else:
-                values = tuple(params.values())
-                values = self._quote_params_for_last_executed_query(values)
-                params = dict(zip(params, values))
-            try:
-                return sql % params
-            except:
-                return sql
-        # For consistency with SQLiteCursorWrapper.execute(), just return sql
-        # when there are no parameters. See #13648 and #17158.
-        else:
-            return sql
-
-    def bulk_insert_sql(self, fields, placeholder_rows):
-        placeholder_rows_sql = (", ".join(row) for row in placeholder_rows)
-        values_sql = ", ".join("(%s)" % sql for sql in placeholder_rows_sql)
-        return "VALUES " + values_sql
